Add notebook, plot scripts and analysis scripts
Those used to be part of the arcus package itself, but I decided
that it was better to split them off in a separate repro.
That way, everything in arcus can be installed with
python setup.py install
and everything in this package is run on the command-line or in the
notebook.